Extraigo funciones para paths de catalogos (y modifico migraciones acorde a los cambios). Creo tempfile con modulo y no manualmente
