+++ infra/apps/catalog/helpers/temp_file_from_url.py
@@ -9,10 +9,7 @@ def temp_file_from_url(url):
     response.raise_for_status()
 
     file_content = response.content
-    name = os.path.basename(response.url)
-    if not name:
-        name = response.url.rsplit('/')[-2]
-    fd = open(f'/{tempfile.gettempdir()}/{name}', 'wb+')
+    fd = tempfile.TemporaryFile()
     fd.write(file_content)
     fd.seek(0)
     return fd
